+++ app/src/lib.rs
@@ -1,15 +1,20 @@
 #![no_std]
 
-pub use iso7816::{Command, Data, Interface, Status};
+pub use iso7816::{command::CommandView, Data, Interface, Status};
 
 pub type Result = iso7816::Result<()>;
 
 /// An App can receive and respond APDUs at behest of the ApduDispatch.
-pub trait App<const C: usize, const R: usize>: iso7816::App {
+pub trait App<const R: usize>: iso7816::App {
     /// Given parsed APDU for select command.
     /// Write response data back to buf, and return length of payload.  Return APDU Error code on error.
     /// Alternatively, the app can defer the response until later by returning it in `poll()`.
-    fn select(&mut self, interface: Interface, apdu: &Command<C>, reply: &mut Data<R>) -> Result;
+    fn select(
+        &mut self,
+        interface: Interface,
+        apdu: CommandView<'_>,
+        reply: &mut Data<R>,
+    ) -> Result;
 
     /// Deselects the app. This is the result of another app getting selected.
     /// App should clear any sensitive state and reset security indicators.
@@ -17,5 +22,5 @@ pub trait App<const C: usize, const R: usize>: iso7816::App {
 
     /// Given parsed APDU for app when selected.
     /// Write response data back to buf, and return length of payload.  Return APDU Error code on error.
-    fn call(&mut self, interface: Interface, apdu: &Command<C>, reply: &mut Data<R>) -> Result;
+    fn call(&mut self, interface: Interface, apdu: CommandView<'_>, reply: &mut Data<R>) -> Result;
 }